+++ CHANGES.md
@@ -2,14 +2,42 @@ Change History
 ==============
 
 
-1.1.0.post1 for MC 1.7.10 (Unreleased)
---------------------------------------
+TODO: 1.3.0 for Minecraft 1.7.10 with Forge 10.13.0 (Unreleased)
+---------------------------------------------------------
+
+New features:
+
+- TODO: Restart prometheus exporter command.
+- TODO: Add player_playtime metric.
+
+
+1.2.1 for Minecraft 1.7.10 with Forge 10.13.0 (Unreleased)
+----------------------------------------------------------
+
+Bug fixes:
+
+- Prevent possible crash on bad start-up.
+
+
+1.2.0 for Minecraft 1.7.10 with Forge 10.13.0 (Unreleased)
+----------------------------------------------------------
+
+New features:
+
+- Added the "collector.mc_dimension_tick_errors" setting to control how to handle inconsistent dimension ticks from misbehaved mods. The new default behavior is to log a debug message rather than crash.
+
+Bug fixes:
+
+- Support inconsistent dimension ticks from misbehaved mods.
+- Support multithreaded dimension ticks.
+
+Miscellaneous:
 
 - Minor documentation.
 
 
-1.1.0 for MC 1.7.10 (2024-04-11)
---------------------------------
+1.1.0 for Minecraft 1.7.10 with Forge 10.13.0 (2024-04-11)
+----------------------------------------------------------
 
 New features:
 
@@ -25,8 +53,8 @@ Miscellaneous:
 - Added "metrics.md".
 
 
-1.0.0 for MC 1.7.10 (2023-12-17)
---------------------------------
+1.0.0 for Minecraft 1.7.10 with Forge 10.13.0 (2023-12-17)
+----------------------------------------------------------
 
 New features: